requirements.txt
aiosqlite>=0.19.0
azure-cosmos>=4.4.0
databases>=0.5.4
jinja2>=3.1.3
jsonschema>=4.20.0
honcho>=1.1.0
msal>=1.25.0
tiktoken>=0.5.0
pyyaml>=6.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.21.0
pytest-json-report>=1.5.0
python-dotenv>=1.0.0
pydantic>=2.6.0
redis>=5.0.0
requests>=2.31.0
setuptools>=65.5.0
streamlit>=1.34.0
termcolor>=2.4.0
imap-tools>=1.10.0
flask>=2.2.0
psycopg2-binary>=2.9.0
ollama>=0.1.0
aiohttp>=3.9.0
alembic>=1.13.0
sqlalchemy>=2.0.0
click>=8.1.0
rich>=13.7.0
prometheus-client>=0.19.0
backoff>=2.2.0
tenacity>=8.2.0
numpy>=1.24.0,<2.0.0
scipy>=1.10.0
httpx>=0.24.0
asyncio-mqtt>=0.13.0
typing-extensions>=4.8.0
transformers>=4.35.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
psutil>=5.9.0
regex>=2022.1.18
filelock>=3.12.0
safetensors>=0.4.0
tokenizers>=0.15.0
packaging>=20.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
jupyter>=1.0.0
jupyterlab>=4.0.0
notebook>=7.0.0
ipywidgets>=8.0.0
ipykernel>=6.25.0
nbformat>=5.9.0
python-multipart>=0.0.6
websockets>=12.0
gunicorn>=21.2.0
starlette>=0.27.0
anyio>=4.0.0
sniffio>=1.3.0
idna>=3.4
certifi>=2023.7.22
charset-normalizer>=3.2.0
urllib3>=1.25.4,<1.27
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.2
pytz>=2023.3
cachetools>=5.3.0
slowapi>=0.1.9
orjson>=3.9.0
pillow>=10.0.0
boto3>=1.26.0
botocore>=1.39.10
watchfiles>=0.21.0
sentence-transformers>=2.2.0
torch>=2.0.0
huggingface-hub>=0.17.0
sentencepiece>=0.1.99
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
